+++ sythil_saas_client/__openerp__.py
@@ -1,6 +1,6 @@
 {
     'name': "Sythil Tech SAAS Client",
-    'version': "1.0.3",
+    'version': "1.1",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",
@@ -10,7 +10,9 @@
     'data': [
         'data/res.users.csv',
         'data/ir.rule.xml',
+        'data/ir.config_parameter.xml',
         'views/ir_actions_todo.xml',
+        'views/res_users_views.xml',
         'views/sythil_saas_client_templates.xml',
     ],
     'demo': [],