+++ sythil_saas_server/models/__init__.py
@@ -1,1 +1,1 @@
-import saas_database, saas_template_database, saas_settings, ir_attachment, saas_modules_builtin, saas_database_domain
+import saas_database, saas_template_database, saas_settings, ir_attachment, saas_modules_builtin, saas_database_domain, payment_acquirer, payment_method